Move builder to a inner class in TwilioVerifySna interface
